requirements.txt
# requirements.txt
omegaconf==2.3.0
numpy==1.26.4
scikit-learn==1.5.0
soundfile==0.12.1
python-json-logger==2.0.7
pyannote.audio==3.3.2
pyannote.metrics==3.2.1
pyannote.core==5.0.0
scipy==1.13.0
requests==2.32.3
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pydantic==2.7.1
httpx==0.27.0
pytest==8.2.0
pytest-asyncio==0.23.6
pytest-cov==5.0.0
huggingface_hub==0.23.4